refactor: 게임 매니저 위치 변경
