+++ src/index.js
@@ -1,4 +1,4 @@
-const GameManager = require('./GameManager');
+const GameManager = require('./controller/GameManager');
 
 const gameManger = new GameManager();
 gameManger.play();
